requirements.txt
absl-py==2.3.1
accelerate==0.23.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
appdirs==1.4.4
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
black==21.4b2
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
chromadb==1.0.20
click==8.2.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle==3.1.1
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
crewai==0.165.1
cryptography==45.0.6
cycler==0.12.1
Cython==3.0.2
datasets==4.1.0
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@a1ce2f956a1d2212ad672e3c47d53405c2fe4312
diffdist==0.1
diffusers==0.35.1
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
dotenv==0.9.9
durationpy==0.10
easydict==1.13
einops==0.8.0
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.116.1
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.7.0
ftfy==6.1.1
functional-cat==0.0.1
future==1.0.0
fvcore==0.1.5.post20221221
gensim==4.3.2
gitdb==4.0.12
GitPython==3.1.45
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.31.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
googletrans==4.0.2
grad-cam==1.5.5
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.1.8
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.17.3
humanfriendly==10.0
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
infinibatch==0.1.1
instructor==1.10.0
iopath==0.1.9
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json-tricks==3.17.3
json5==0.12.1
json_repair==0.25.2
jsonpickle==4.1.1
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
kornia==0.7.0
kornia_rs==0.1.9
kubernetes==33.1.0
lazy_loader==0.4
litellm==1.74.9
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
maskrcnn-benchmark @ git+https://github.com/microsoft/GLIP.git@9dda9558c1ef59bb6cdc8e896e2bcab775a68ff0
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
MedCLIP==0.0.3
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
mup==1.0.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.8.1
numpy==1.26.4
oauthlib==3.3.1
omegaconf==2.3.0
onnxruntime==1.22.0
open_clip_torch==2.26.1
openai==1.99.9
opencv-python==4.8.1.78
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.2
overrides==7.7.0
packaging==25.0
pandas==2.0.3
parso==0.8.5
pathspec==0.12.1
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
Pillow==9.4.0
platformdirs==4.3.8
portalocker==2.7.0
posthog==5.4.0
prettytable==3.16.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycocotools==2.0.7
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git@2723c7fa4f2ead16ae1ce3d8afe977724bb8f87f
pydot==4.0.1
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.14.1
pyparsing==3.2.3
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytorch-gradcam==0.2.1
pytz==2025.2
pyvis==0.3.2
PyWavelets==1.9.0
PyYAML==6.0.1
pyzmq==27.0.2
referencing==0.36.2
regex==2023.10.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
safetensors==0.6.2
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.16.1
seaborn==0.13.2
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
sentencepiece==0.1.99
sentry-sdk==2.35.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.47.3
supervision==0.26.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
termcolor==3.1.0
textaugment==2.0.0
textblob==0.19.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.9.20
tiktoken==0.11.0
timm==1.0.20
tokenizers==0.14.1
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.34.0
ttach==0.0.3
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.184
ultralytics-thop==2.0.16
uritemplate==4.2.0
urllib3==2.5.0
uv==0.8.13
uvicorn==0.35.0
uvloop==0.21.0
validators==0.35.0
vision-datasets==0.2.2
wandb==0.21.1
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wget==3.2
wrapt==1.17.3
xxhash==3.5.0
yacs==0.1.8
yapf==0.43.0
yarl==1.20.1
zipp==3.23.0
